requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Werkzeug
requests
marshmallow
pytest
waitress
